# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "mappa-api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Guionardo Furlan",
            "email": "guionardo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.4.0"
    },
    "basePath": "/",
    "paths": {
...
        "/mappa/conquistas/{cod_secao}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mappa-proxy"
                ],
                "summary": "Lista de conquistas da secão",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Código Seção",
                        "name": "cod_secao",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MappaMarcacaoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    }
                }
            }
        },
        "/mappa/escotista/{userId}": {
            "get": {
                "description": "Informações do escotista, associado e grupo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mappa-proxy"
                ],
                "summary": "Detalhes do escotista",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MappaDetalhesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    }
                }
            }
        },
        "/mappa/escotista/{userId}/secoes": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mappa-proxy"
                ],
                "summary": "Seções do escotista",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MappaSecaoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    }
                }
            }
        },
        "/mappa/especialidades": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "db"
                ],
                "summary": "Lista de especialidades e items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MappaEspecialidadeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    }
                }
            }
        },
        "/mappa/login": {
            "post": {
                "description": "User login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mappa-proxy"
                ],
                "summary": "Mappa Login handler",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MappaLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    }
                }
            }
        },
        "/mappa/marcacoes/{cod_secao}": {
            "get": {
                "description": "Lista de marcações da sessão",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mappa-proxy"
                ],
                "summary": "MappaMarcacoes handler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Código Seção",
                        "name": "cod_secao",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.MappaMarcacaoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    }
                }
            }
        },
        "/mappa/progressoes/{ramo}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "db"
                ],
                "summary": "Lista de progressões do ramo",
                "parameters": [
                    {
                        "enum": [
                            "L",
                            "E",
                            "S",
                            "P"
                        ],
                        "type": "string",
                        "description": "Ramo",
                        "name": "ramo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.MappaProgressaoResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ReplyMessage": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "requests.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "responses.MappaAssociadoResponse": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "integer"
                },
                "codigoCategoria": {
                    "type": "integer"
                },
                "codigoEquipe": {
                    "type": "integer"
                },
                "codigoFoto": {
                    "type": "integer"
                },
                "codigoRamo": {
                    "type": "integer"
                },
                "codigoRamoAdulto": {
                    "type": "integer"
                },
                "codigoSegundaCategoria": {
                    "type": "integer"
                },
                "codigoTerceiraCategoria": {
                    "type": "integer"
                },
                "dataAcompanhamento": {
                    "type": "string"
                },
                "dataNascimento": {
                    "type": "string"
                },
                "dataValidade": {
                    "type": "string"
                },
                "linhaFormacao": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "nomeAbreviado": {
                    "type": "string"
                },
                "numeroDigito": {
                    "type": "integer"
                },
                "sexo": {
                    "type": "string"
                },
                "username": {
                    "type": "integer"
                }
            }
        },
        "responses.MappaDetalhesResponse": {
            "type": "object",
            "properties": {
                "associado": {
                    "$ref": "#/definitions/responses.MappaAssociadoResponse"
                },
                "escotista": {
                    "$ref": "#/definitions/responses.MappaEscotistaResponse"
                },
                "grupos": {
                    "$ref": "#/definitions/responses.MappaGrupoResponse"
                }
            }
        },
        "responses.MappaEscotistaResponse": {
            "type": "object",
            "properties": {
                "ativo": {
                    "$ref": "#/definitions/types.Bool"
                },
                "codigo": {
                    "type": "integer"
                },
                "codigoAssociado": {
                    "type": "integer"
                },
                "codigoFoto": {
                    "type": "integer"
                },
                "codigoGrupo": {
                    "type": "integer"
                },
                "codigoRegiao": {
                    "type": "string"
                },
                "nomeCompleto": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "responses.MappaEspecialidadeItemResponse": {
            "type": "object",
            "properties": {
                "codigoEspecialidade": {
                    "description": "Id                  int    `json:\"id\"`",
                    "type": "integer"
                },
                "descricao": {
                    "type": "string"
                },
                "numero": {
                    "type": "integer"
                }
            }
        },
        "responses.MappaEspecialidadeResponse": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "integer"
                },
                "descricao": {
                    "type": "string"
                },
                "itens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.MappaEspecialidadeItemResponse"
                    }
                },
                "prerequisito": {
                    "type": "string"
                },
                "ramoConhecimento": {
                    "type": "string"
                }
            }
        },
        "responses.MappaGrupoResponse": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "integer"
                },
                "codigoModalidade": {
                    "type": "integer"
                },
                "codigoRegiao": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "responses.MappaLoginResponse": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ttl": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "responses.MappaMarcacaoResponse": {
            "type": "object",
            "properties": {
                "codigoAssociado": {
                    "type": "integer"
                },
                "codigoAtividade": {
                    "type": "integer"
                },
                "codigoUltimoEscotista": {
                    "type": "integer"
                },
                "dataAtividade": {
                    "type": "string"
                },
                "dataHoraAtualizacao": {
                    "type": "string"
                },
                "dataStatusEscotista": {
                    "type": "string"
                },
                "segmento": {
                    "type": "string"
                }
            }
        },
        "responses.MappaProgressaoResponse": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "integer"
                },
                "codigoCaminho": {
                    "type": "integer"
                },
                "codigoCompetencia": {
                    "type": "integer"
                },
                "codigoDesenvolvimento": {
                    "type": "integer"
                },
                "codigoRegiao": {
                    "type": "string"
                },
                "codigoUeb": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "numeroGrupo": {
                    "type": "integer"
                },
                "ordenacao": {
                    "type": "integer"
                },
                "segmento": {
                    "type": "string"
                }
            }
        },
        "responses.MappaSecaoResponse": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "integer"
                },
                "codigoGrupo": {
                    "type": "integer"
                },
                "codigoRegiao": {
                    "type": "string"
                },
                "codigoTipoSecao": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "subsecoes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.MappaSubSecaoResponse"
                    }
                }
            }
        },
...
        "responses.MappaSubSecaoResponse": {
            "type": "object",
            "properties": {
                "associados": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.MappaAssociadoResponse"
                    }
                },
                "codigo": {
                    "type": "integer"
                },
                "codigoLider": {
                    "type": "integer"
                },
                "codigoSecao": {
                    "type": "integer"
                },
                "codigoViceLider": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "types.Bool": {
            "type": "object"
        }
    }
}